:recycle: Refactor(color): apply norm && divide to multiply
